third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/serviceworker/chromium.performance-timeline-expected.txt

This is a testharness.js-based test.
PASS Test Performance Timeline API in Service Worker 
PASS User Timing 
FAIL Resource Timing assert_equals: expected 2 but got 1
Harness: the test ran to completion.
